# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassifyInput {
  /** User emergency input text */
  transcript: string;
}

export type ClassificationResultUrgency = typeof ClassificationResultUrgency[keyof typeof ClassificationResultUrgency];


export const ClassificationResultUrgency = {
  CRITICAL: 'CRITICAL',
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export type ClassificationResultConfidence = typeof ClassificationResultConfidence[keyof typeof ClassificationResultConfidence];


export const ClassificationResultConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Helpline {
  name: string;
  number: string;
  when: string;
}

export interface ClassificationResult {
  category: string;
  subcategory: string;
  urgency: ClassificationResultUrgency;
  confidence: ClassificationResultConfidence;
  clarificationNeeded: boolean;
  /** @nullable */
  question?: string | null;
  helplines?: Helpline[];
  immediateActions?: string[];
}

export type ConversationTurnRole = typeof ConversationTurnRole[keyof typeof ConversationTurnRole];


export const ConversationTurnRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ConversationTurn {
  role: ConversationTurnRole;
  content: string;
}

export interface GuidanceInput {
  transcript: string;
  history?: ConversationTurn[];
}

export interface TtsInput {
  text: string;
}

export interface TtsOutput {
  audio: string;
  format: string;
  text?: string;
}