# lib/api-zod/src/generated/types/classificationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { ClassificationResultConfidence } from './classificationResultConfidence';
import type { ClassificationResultUrgency } from './classificationResultUrgency';
import type { Helpline } from './helpline';

export interface ClassificationResult {
  category: string;
  subcategory: string;
  urgency: ClassificationResultUrgency;
  confidence: ClassificationResultConfidence;
  clarificationNeeded: boolean;
  /** @nullable */
  question?: string | null;
  helplines?: Helpline[];
  immediateActions?: string[];
}