# lib/api-zod/src/generated/types/classifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassifyInput {
  /** User emergency input text */
  transcript: string;
}