# lib/api-zod/src/generated/types/conversationTurn.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationTurnRole } from './conversationTurnRole';

export interface ConversationTurn {
  role: ConversationTurnRole;
  content: string;
}