# lib/api-zod/src/generated/types/helpline.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface Helpline {
  name: string;
  number: string;
  when: string;
}