# lib/api-zod/src/generated/types/ttsOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface TtsOutput {
  audio: string;
  format: string;
  text?: string;
}